+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## 0.4.0
 
+- add functions:
+  - `cleanupHTTPHeaders()`
+  - `rad2deg()`
+  - `toJSONValue()`
+- improve typings of:
+  - `isNil()`
+- **BREAKING CHANGE**: fix typos like `WillNullableProps` to `WithNullableProps`
 - require at least [Node 14](https://github.com/nodejs/node/blob/master/doc/changelogs/CHANGELOG_V14.md) to build now
 - using new [ESLint settings](https://github.com/egomobile/eslint-config-ego)
 - `npm update`s